+++ src/main/java/dev/aika/chatjs/server/ServerEvents.java
@@ -4,9 +4,12 @@ import dev.aika.chatjs.ChatJS;
 import dev.aika.chatjs.ChatJSUtil;
 import dev.aika.chatjs.command.ChatJSCommands;
 import dev.aika.chatjs.kubejs.OpenAIWrapper;
+import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.level.storage.LevelResource;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.neoforge.event.RegisterCommandsEvent;
+import net.neoforged.neoforge.event.entity.player.PlayerEvent;
 import net.neoforged.neoforge.event.server.ServerStartingEvent;
 import net.neoforged.neoforge.event.server.ServerStoppingEvent;
 import org.slf4j.Logger;
@@ -15,6 +18,7 @@ import org.slf4j.MarkerFactory;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Objects;
 
 public final class ServerEvents {
     private static final Logger log = ChatJS.LOGGER;
@@ -50,6 +54,18 @@ public final class ServerEvents {
             OpenAIWrapper.setApiKey(null);
         } catch (IOException e) {
             log.error(marker, "Failed to save secret file", e);
+        }
+    }
+
+    @SubscribeEvent
+    public void onPlayerLoggedIn(PlayerEvent.PlayerLoggedInEvent event) {
+        if (event.getEntity() instanceof ServerPlayer player) {
+            if ((boolean) ChatJS.CONFIG.get("apikey_not_set_warning") && SecretManager.INSTANCE.getProperty("apikey") == null) {
+                // Check player permissions
+                if (Objects.requireNonNull(player.getServer()).isSingleplayer() || player.hasPermissions(2)) {
+                    player.sendSystemMessage(Component.translatable("message.chatjs.apikey_not_set_warning"));
+                }
+            }
         }
     }
 }